+++ cmd/substreams-sink-kv/serve.go
@@ -30,7 +30,6 @@ var serveCmd = Command(serveRunE,
 	Flags(func(flags *pflag.FlagSet) {
 		flags.String("listen-addr", ":7878", "Listen via GRPC Connect-Web on this address")
 		flags.Bool("listen-ssl-self-signed", false, "Listen with an HTTPS server (with self-signed certificate)")
-		flags.String("api-prefix", "", "Prefix that will be added to the Connect Web routes")
 	}),
 	Description(`
 		Test
@@ -128,7 +127,7 @@ func setupServer(cmd *cobra.Command, pkg *pbsubstreams.Package, kvDB *db.DB) (se
 			return nil, fmt.Errorf("find proto file descriptor: %w", err)
 		}
 
-		config, err := wasmquery.NewServiceConfig(fileDesc, wasmServ.GrpcService, sflags.MustGetString(cmd, "api-prefix"))
+		config, err := wasmquery.NewServiceConfig(fileDesc, wasmServ.GrpcService, sflags.MustGetString(cmd, "server-api-prefix"))
 		if err != nil {
 			return nil, fmt.Errorf("failed to setup grpc config: %w", err)
 		}