+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v2.1.2
+
+### Changed
+
+* _Deprecation_ The flag `substreams-sink-kv inject --listen-addr=...` must now be used like `substreams-sink-kv inject --server-listen-addr=...`
+
+### Fixed
+
+* Added back missing flag `--server-api-prefix` to define API prefix when server mode is active.
+
 ## v2.1.1
 
 ### Added